# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventario Ganaderia API
  version: 1.1.0
  description: API REST para trazabilidad sanitaria, inventario y operaciones ganaderas.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and issue access/refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                tenantId:
                  type: string
                  format: uuid
              required: [email, password]
      responses:
        '200':
          description: Tokens issued
        '401':
          description: Invalid credentials
  /auth/register:
    post:
      tags: [Auth]
      summary: Register tenant and first admin user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                tenantName:
                  type: string
                registrationCode:
                  type: string
      responses:
        '201':
          description: Account created
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token (refresh token from body or HttpOnly cookie)
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Access token refreshed and refresh token rotated
          headers:
            Set-Cookie:
              description: Rotated HttpOnly refresh token cookie
              schema:
                type: string
        '401':
          description: Invalid, expired, revoked token or inactive user
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token (body or HttpOnly cookie)
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token revoked
        '401':
          description: Invalid or expired token
  /auth/switch-tenant:
    post:
      tags: [Auth]
      summary: Switch active tenant and issue new tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId:
                  type: string
                  format: uuid
              required: [tenantId]
      responses:
        '200':
          description: Tenant switched
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user profile
      responses:
        '200':
          description: Current user
        '404':
          description: User not found
...
  /treatments/{id}/close:
    post:
      tags: [Treatments]
      summary: Close treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Treatment closed
  /treatments/by-animal/{animalId}:
    get:
      tags: [Treatments]
      summary: List treatments by animal
      parameters:
        - name: animalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Treatment list by animal
...